Initial Commit
These are old files, part of an experiment. They need cleaning and
commenting. There are too many references for me to remember after so
long, but I will try to locate and list them, when I get time.

This project currently does not compile, I am in the process of updating
it to use node 0.12.4. Until then the source is good for reference, but
not running.